+++ public/atom
@@ -4,6 +4,21 @@
     <subtitle>Blog</subtitle>
     <link href="/atom" rel="self" type="application/rss+xml"/>
     <link href="/" />
-    <updated>2021-01-21T15:32:37.828Z</updated>
+    <updated>2021-01-22T15:57:27.122Z</updated>
     <id>My Notion Blog</id>
+    <entry>
+      <id>/start-table-tennis-blog</id>
+      <title>Table Tennis Blog始めます</title>
+      <link href="/start-table-tennis-blog"/>
+      <updated>2021-01-22T15:00:00.529Z</updated>
+      <content type="xhtml">
+        <div xmlns="http://www.w3.org/1999/xhtml">
+          
+          <p class="more">
+            <a href="/start-table-tennis-blog">Read more</a>
+          </p>
+        </div>
+      </content>
+      
+    </entry>
   </feed>
