+++ src/components/header.tsx
@@ -6,7 +6,6 @@ import styles from '../styles/header.module.css'
 
 const navItems: { label: string; page?: string; link?: string }[] = [
   { label: 'Home', page: '/' },
-  { label: 'Blog', page: '/blog' },
   { label: 'Contact', page: '/contact' },
 ]
