+++ src/pages/index.tsx
@@ -3,15 +3,92 @@ import Header from '../components/header'
 import ExtLink from '../components/ext-link'
 import Features from '../components/features'
 import GitHub from '../components/svgs/github'
+
+import blogStyles from '../styles/blog.module.css'
 import sharedStyles from '../styles/shared.module.css'
 
-export default () => (
-  <>
-    <Header titlePre="Home" />
-    <div className={sharedStyles.layout}>
-      <img src="/profile.png" height="85" width="250" alt="Profile" />
-      <h1>Ryoichi Table Tennis Blog</h1>
-      <h2>10年ブランクでどれだけ強くなれるのか</h2>
-    </div>
-  </>
-)
+import { getBlogLink, getDateStr, postIsPublished } from '../lib/blog-helpers'
+import { textBlock } from '../lib/notion/renderers'
+import getNotionUsers from '../lib/notion/getNotionUsers'
+import getBlogIndex from '../lib/notion/getBlogIndex'
+
+export async function getStaticProps({ preview }) {
+  const postsTable = await getBlogIndex()
+
+  const authorsToGet: Set<string> = new Set()
+  const posts: any[] = Object.keys(postsTable)
+    .map(slug => {
+      const post = postsTable[slug]
+      // remove draft posts in production
+      if (!preview && !postIsPublished(post)) {
+        return null
+      }
+      post.Authors = post.Authors || []
+      for (const author of post.Authors) {
+        authorsToGet.add(author)
+      }
+      return post
+    })
+    .filter(Boolean)
+
+  const { users } = await getNotionUsers([...authorsToGet])
+
+  posts.map(post => {
+    post.Authors = post.Authors.map(id => users[id].full_name)
+  })
+
+  return {
+    props: {
+      preview: preview || false,
+      posts,
+    },
+    unstable_revalidate: 10,
+  }
+}
+
+export default ({ posts = [], preview }) => {
+  return (
+    <>
+      <Header titlePre="Blog" />
+
+      {preview && (
+        <div className={blogStyles.previewAlertContainer}>
+          <div className={blogStyles.previewAlert}>
+            <b>Note:</b>
+            {` `}Viewing in preview mode{' '}
+            <Link href={`/api/clear-preview`}>
+              <button className={blogStyles.escapePreview}>Exit Preview</button>
+            </Link>
+          </div>
+        </div>
+      )}
+      <div className={`${sharedStyles.layout} ${blogStyles.blogIndex}`}>
+        <img src="/profile.png" height="85" width="250" alt="Profile" />
+        <h1>Ryoichi Table Tennis Blog</h1>
+        <h2>10年ブランクでどれだけ強くなれるのか</h2>
+        {posts.length === 0 && (
+          <p className={blogStyles.noPosts}>There are no posts yet</p>
+        )}
+        {posts.map(post => {
+          return (
+            <div className={blogStyles.postPreview} key={post.Slug}>
+              <h3>
+                <Link href="/[slug]" as={getBlogLink(post.Slug)}>
+                  <div className={blogStyles.titleContainer}>
+                    {!post.Published && (
+                      <span className={blogStyles.draftBadge}>Draft</span>
+                    )}
+                    <a>{post.Page}</a>
+                  </div>
+                </Link>
+              </h3>
+              {post.Date && (
+                <div className="posted">Posted: {getDateStr(post.Date)}</div>
+              )}
+            </div>
+          )
+        })}
+      </div>
+    </>
+  )
+}
